+++ app/controllers/users_controller.rb
@@ -22,7 +22,7 @@ class UsersController < ApplicationController
   end
 
   def show
-    @user = User.where('lower(user_name) = ?', (params[:user_id].presence || params[:id]).downcase).first
+    @user = User.find_by('lower(user_name) = ?', (params[:user_id].presence || params[:id]).downcase)
 
     return render 'not_found', status: :not_found if @user.nil?
 
@@ -36,62 +36,67 @@ class UsersController < ApplicationController
     if params[:setup]
       setup = @user.setups.find(params[:setup])
       all_possessions = setup.possessions
-                             .joins(:product)
-                             .includes([product: [{ sub_categories: :category }, :brand]])
-                             .order('LOWER(products.name)')
-                             .map { |bookmark| ItemPresenter.new(bookmark) }
+      all_custom_products = setup.possessions
     else
       all_possessions = @user.possessions
-                             .joins(:product)
-                             .includes([product: [{ sub_categories: :category }, :brand]])
-                             .order('LOWER(products.name)')
-                             .map { |bookmark| ItemPresenter.new(bookmark) }
+      all_custom_products = @user.possessions
     end
+
+    all_possessions = all_possessions.joins(:product)
+                                     .includes([product: [{ sub_categories: :category }, :brand]])
+                                     .order('LOWER(products.name)')
+                                     .map { |possession| ItemPresenter.new(possession) }
+    all_custom_products = all_custom_products.joins(:custom_product)
+                                             .includes([custom_product: [{ sub_categories: :category }]])
+                                             .map { |possession| CustomProductPresenter.new(possession) }
+    all = (all_possessions + all_custom_products).sort_by(&:short_name)
 
     if params[:category].present?
       @sub_category = SubCategory.friendly.find(params[:category])
-      @possessions = all_possessions.select { |possession| @sub_category.products.include?(possession.product) }
+      @possessions = all.select do |possession|
+        @sub_category.products.include?(possession.product) ||
+          @sub_category.custom_products.include?(possession.custom_product)
+      end
     else
-      @possessions = all_possessions
+      @possessions = all
     end
 
-    @categories = all_possessions.map(&:product)
-                                 .flat_map(&:sub_categories)
-                                 .sort_by(&:name)
-                                 .uniq
-                                 .group_by(&:category)
-                                 .sort_by { |category| category[0].order }
-                                 # rubocop:disable Style/BlockDelimiters
-                                 .map { |c|
-                                   [
-                                     c[0],
-                                     c[1].map { |sub_category|
-                                       # rubocop:enable Style/BlockDelimiters
-                                       {
-                                         name: sub_category.name,
-                                         friendly_id: sub_category.friendly_id,
-                                         path: (
-                                           if setup
-                                             user_setup_path(
-                                               user_id: @user.user_name.downcase,
-                                               category: sub_category.friendly_id,
-                                               setup: setup.id
-                                             )
-                                           else
-                                             user_path(id: @user.user_name.downcase, category: sub_category.friendly_id)
-                                           end
-                                         )
-                                       }
-                                     }
-                                   ]
-                                 }
+    @categories = all.flat_map(&:sub_categories)
+                     .sort_by(&:name)
+                     .uniq
+                     .group_by(&:category)
+                     .sort_by { |category| category[0].order }
+                     # rubocop:disable Style/BlockDelimiters
+                     .map { |c|
+                       [
+                         c[0],
+                         c[1].map { |sub_category|
+                           # rubocop:enable Style/BlockDelimiters
+                           {
+                             name: sub_category.name,
+                             friendly_id: sub_category.friendly_id,
+                             path: (
+                               if setup
+                                 user_setup_path(
+                                   user_id: @user.user_name.downcase,
+                                   category: sub_category.friendly_id,
+                                   setup: setup.id
+                                 )
+                               else
+                                 user_path(id: @user.user_name.downcase, category: sub_category.friendly_id)
+                               end
+                             )
+                           }
+                         }
+                       ]
+                     }
     @reset_path = @user.profile_path
 
     render 'show', locals: { include_images: true }
   end
 
   def prev_owneds
-    @user = User.where('lower(user_name) = ?', (params[:user_id].presence || params[:id]).downcase).first
+    @user = User.find_by('lower(user_name) = ?', (params[:user_id].presence || params[:id]).downcase)
 
     return render 'not_found', status: :not_found if @user.nil?