+++ app/models/user.rb
@@ -18,6 +18,7 @@ class User < ApplicationRecord
   has_many :setup_possessions, through: :setups
   has_many :bookmarks, dependent: :destroy
   has_many :prev_owneds, dependent: :destroy
+  has_many :custom_products, dependent: :destroy, through: :possessions
   has_one_attached :avatar do |attachable|
     attachable.variant :thumb, resize_to_limit: [320, 320], format: :webp
     attachable.variant :thumb_avif, resize_to_limit: [320, 320], format: :avif