+++ app/presenters/item_presenter.rb
@@ -5,6 +5,8 @@ class ItemPresenter
 
   attr_reader :object, :product, :product_variant
 
+  delegate :name, to: :brand, prefix: true
+
   def initialize(object)
     @object = object
     @product = object.product
@@ -23,9 +25,29 @@ class ItemPresenter
     @product.display_name
   end
 
-  def product_or_variant_path
+  def show_path
     return @product_variant.path if @product_variant.present?
 
     product_path(id: @product.friendly_id)
+  end
+
+  def edit_path
+    if @product_variant.present?
+      return product_edit_variant_path(product_id: @product.friendly_id, variant: @product_variant.friendly_id)
+    end
+
+    edit_product_path(id: @product.friendly_id)
+  end
+
+  def brand
+    @product.brand
+  end
+
+  def categories
+    @product.categories
+  end
+
+  def sub_categories
+    @product.sub_categories
   end
 end