+++ db/migrate/20240314153138_add_user_id_to_custom_product.rb
@@ -1,0 +1,6 @@
+class AddUserIdToCustomProduct < ActiveRecord::Migration[7.1]
+  def change
+    add_column :custom_products, :user_id, :bigint, null: false
+    change_column_null :custom_products, :name, false
+  end
+end