+++ db/migrate/20240314153951_add_custom_products_sub_categories_join_table.rb
@@ -1,0 +1,5 @@
+class AddCustomProductsSubCategoriesJoinTable < ActiveRecord::Migration[7.1]
+  def change
+    create_join_table :custom_products, :sub_categories
+  end
+end