+++ db/migrate/20240314160029_add_custom_products_id_to_possessions.rb
@@ -1,0 +1,7 @@
+class AddCustomProductsIdToPossessions < ActiveRecord::Migration[7.1]
+  def change
+    add_column :possessions, :custom_product_id, :bigint
+    change_column_null :possessions, :product_id, true
+    remove_column :custom_products, :user_id
+  end
+end